+++ test/Integration/Dialect/TOSA/i16xi16_mul_elem_16x1024_broadcast_1024/i16xi16_mul_elem_2d_broadcast_1d.mlir
@@ -14,7 +14,7 @@
 
 module {
   func.func @dut(%arg0: tensor<16x1024xi16>, %arg1: tensor<1024xi16>) -> (tensor<16x1024xi16>) {
-    %1 = "tosa.mul"(%arg0,%arg1) {shift = 0 : i32} : (tensor<16x1024xi16>, tensor<1024xi16>)  -> (tensor<16x1024xi16>)
+    %1 = "tosa.mul"(%arg0,%arg1) {shift = 0 : i8} : (tensor<16x1024xi16>, tensor<1024xi16>)  -> (tensor<16x1024xi16>)
     return %1 : tensor<16x1024xi16>
   }
 }
